[ADD] device association, constants
Added the table names to the constants file for easier usage in the sqlite service file(to be renamed).
Added the full function for assigning a device to a client.
Tables will be migrated to a single db file due to improved understanding of the sqllite functioning.
Added date and time package for getting the current date and time.
Fixed some of the JSON model files.